use matchPath in whenActive
Instead of relying on a <Route> component, switch to calling matchPath directly.

This commit also adds support for an isActive option. This should be a function that takes two arguments: pathname and props. pathname is the current pathname and props is the props passed to the wrapped component. The function should use these values to determine whether the component is active or not (and return the proper boolean)